fix: TUI scroll safety, query robustness, and truncation zero-copy
Analytics and explorer detail panes now track last_detail_max_scroll via
Cell to prevent scrolling past content bounds — eliminates blank-screen
state when rapidly scrolling long insight cards or message details.

- queries.rs: Switch fetch_unacked_for_agent from get_named() to
  positional get_as() for resilience against column name changes
- query_integration.rs: Fix block_on helper to use runtime-backed Cx
  via rt.handle().spawn() instead of Cx::for_testing() which lacks
  io_driver/timer_driver backing (root cause of test hangs)
- tui_widgets.rs: Return Cow<str> from truncate_width to avoid
  unnecessary String allocation when input fits within max_width
- contacts.rs: Fix integer overflow in scaled_level range calculation
  (i64::from(max - min) could overflow when min is negative)
- dashboard.rs: Adapt truncate() return type to match Cow signature
- messages.rs: Call .into_owned() on Cow return from truncate_width
- build_slots.rs: Remove incorrect `if t > 0` guard on TTL matching —
  clamp(60, ...) already handles the lower bound correctly, and the
  guard silently fell through to the default for explicit zero values

+++ crates/mcp-agent-mail-server/src/tui_screens/contacts.rs
@@ -1345,7 +1345,7 @@ fn scaled_level(value: u32, max_value: u32, min: i32, max: i32) -> i32 {
         return min;
     }
     let clamped = value.min(max_value);
-    let range = i64::from(max - min);
+    let range = i64::from(max) - i64::from(min);
     let scaled = i64::from(min)
         + ((i64::from(clamped) * range) + i64::from(max_value / 2)) / i64::from(max_value);
     i32::try_from(scaled.clamp(i64::from(min), i64::from(max))).unwrap_or(max)
